@@ -1,5 +1,5 @@
 name: fake_http_client
-version: 0.0.3
+version: 0.2.0
 description: >-
   A package for faking Dart/Flutter HttpClient's responses.
   It is handy for integration tests. Also it may help a lot during 
@@ -7,8 +7,8 @@ description: >-
 homepage: https://github.com/Pavel-Sulimau/fake_http_client
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  test: ^1.15.7
-  http: ^0.12.2
+  http: ^0.13.0
+  test: any
